moved some functions to helper files and fixed a ton of PEP-8 violations.
I had to add bot as a parameter to is_owner since I moved it. It might be better to just make this a class method. I also added ffmpeg to the gitignore so I can click stage all at long last.
