+++ .gitignore
@@ -113,3 +113,4 @@ Contact GitHub API Training Shop Blog About
 .idea/
 settings.json
 Databases/
+ffmpeg.exe
