+++ RyuZU/Memes/__init__.py
@@ -1,20 +1,10 @@
-import datetime
 from random import choice
 
 import requests
-from RyuZU.Core import Cog
 from discord.ext import commands
 
-
-def prune_respects(db):
-    now = datetime.datetime.now()
-    all = db.all()
-    d = []
-    for a in all:
-        dt = datetime.datetime.strptime(a['timestamp'], '%b %d %Y %I:%M%p')
-        if now - dt > datetime.timedelta(days=1):
-            d.append(a.eid)
-    db.remove(eids=d)
+from RyuZU.Core import Cog
+from .helpers import prune_respects
 
 
 class Memes(Cog):
