+++ RyuZU/MusicPlayer/__init__.py
@@ -3,8 +3,9 @@ import math
 import platform
 
 import discord
-from RyuZU.Core import Cog
 from discord.ext import commands
+
+from RyuZU.Core import Cog
 
 if not discord.opus.is_loaded() and platform.system() == "Windows":
     discord.opus.load_opus('opus')
