+++ RyuZU/Util/__init__.py
@@ -1,7 +1,8 @@
 import random
 
-from RyuZU.Core import Cog
 from discord.ext import commands
+
+from RyuZU.Core import Cog
 
 
 class Util(Cog):
