+++ RyuZU/__init__.py
@@ -36,7 +36,7 @@ async def on_ready():
 @bot.command(pass_context=True, hidden=True)
 async def load(ctx, extension_name: str):
     """Loads an extension."""
-    if not is_owner(ctx.message.author):
+    if not is_owner(bot, ctx.message.author):
         await bot.say("You must be {0}'s owner to do this.".format(bot.user.name))
         return
     try:
@@ -50,7 +50,7 @@ async def load(ctx, extension_name: str):
 @bot.command(pass_context=True, hidden=True)
 async def unload(ctx, extension_name: str):
     """Unloads an extension."""
-    if not is_owner(ctx.message.author):
+    if not is_owner(bot, ctx.message.author):
         await bot.say("You must be {0}'s owner to do this.".format(bot.user.name))
         return
     bot.unload_extension(extension_name)
@@ -60,7 +60,7 @@ async def unload(ctx, extension_name: str):
 @bot.command(pass_context=True, hidden=True)
 async def shutdown(ctx):
     """Shuts down the bot"""
-    if not is_owner(ctx.message.author):
+    if not is_owner(bot, ctx.message.author):
         await bot.say("You must be {0}'s owner to do this.".format(bot.user.name))
         return
     await bot.change_presence(game=discord.Game(name="Shutting down"))
